# 8145948: C2: Initializing volatile fields to default values should be optimized