WIP.

Additional testing:
 - [x] New IR tests pass
 - [ ] Linux x86_64 server fastdebug, `all`
 - [ ] Linux AArch64 server fastdebug, `all`

```
Benchmark                                        Mode  Cnt   Score   Error  Units

# Baseline
ConstructorDefaultInitBarriers.plain_init_0      avgt    3   5.815 ± 0.187  ns/op
ConstructorDefaultInitBarriers.plain_init_42     avgt    3   5.511 ± 0.053  ns/op
ConstructorDefaultInitBarriers.volatile_init_0   avgt    3   5.737 ± 0.079  ns/op
ConstructorDefaultInitBarriers.volatile_init_42  avgt    3  11.885 ± 0.950  ns/op

# Patched
ConstructorDefaultInitBarriers.plain_init_0      avgt    3   5.820 ± 0.161  ns/op
ConstructorDefaultInitBarriers.plain_init_42     avgt    3   5.504 ± 0.108  ns/op
ConstructorDefaultInitBarriers.volatile_init_0   avgt    3  11.969 ± 0.223  ns/op  ; <----
ConstructorDefaultInitBarriers.volatile_init_42  avgt    3  12.028 ± 1.671  ns/op
```